+++ Extensions/Git.Branch.UGit.Extension.ps1
@@ -5,6 +5,11 @@
     Outputs git branch as objects (unless, -m, -c, -column, -format, or -show-current are passed)
 .EXAMPLE
     git branch
+.EXAMPLE
+    git branch |                                          # Get all branches
+        Where-Object -Not IsCurrentBranch |               # where it is not the current branch
+        Where-Object BranchName -NotIn 'main', 'master' | # and the name is not either main or master
+        git branch -d                                     # then attempt to delete the branch.
 #>
 # It's an extension for Out-Git
 [Management.Automation.Cmdlet("Out","Git")]